Add an image vector layer for rendering vectors to an image
