+++ examples/image-vector-layer.html
@@ -1,9 +1,9 @@
 ---
 layout: example.html
-title: Image Vector Layer
-shortdesc: Example of an image vector layer.
+title: Vector Image Layer
+shortdesc: Example of rendering vector data as an image layer.
 docs: >
-  <p>This example uses <code>ol/layer/Vector</code> with `renderMode: 'image'`. This mode results in faster rendering during interaction and animations, at the cost of less accurate rendering.</p>
+  <p>This example uses <code>ol/layer/VectorImage</code> for faster rendering during interaction and animations, at the cost of less accurate rendering.</p>
 tags: "vector, image"
 ---
 <div id="map" class="map"></div>
