+++ src/ol/layer.js
@@ -9,4 +9,5 @@ export {default as Image} from './layer/Image.js';
 export {default as Layer} from './layer/Layer.js';
 export {default as Tile} from './layer/Tile.js';
 export {default as Vector} from './layer/Vector.js';
+export {default as VectorImage} from './layer/VectorImage.js';
 export {default as VectorTile} from './layer/VectorTile.js';
